port to Opee, fix for el cap

+++ FlashlightApp/FlashlightKit/FlashlightKit/TimeParser.c
@@ -364,7 +364,7 @@ static void reverse(char s[])
     int i, j;
     char c;
     
-    for (i = 0, j = strlen(s)-1; i<j; i++, j--) {
+    for (i = 0, j = (int)strlen(s)-1; i<j; i++, j--) {
         c = s[i];
         s[i] = s[j];
         s[j] = c;
